DD_Chat_Bot/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-core
langchain-chroma
langchain-groq
langgraph
openai
chromadb
pydantic
feedparser
groq
python-dotenv
sentence-transformers
schedule
requests
yfinance
pandas
numpy
marketaux
newspaper3k
lxml_html_clean
twilio
python-multipart